+++ phpBB2/includes/page_header.php
@@ -618,9 +618,11 @@ if ( $marknow == 'globmsg' && $userdata['session_logged_in'] )
 
 if ( $userdata['ct_global_msg_read'] == 1 && $userdata['session_logged_in'] && $ctracker_settings['global_message'] != '' )
 {
-	$global_message_output = ($ctracker_settings['global_message_type'] == 1)
-		? $ctracker_settings['global_message']
-		: sprintf($lang['ctracker_gmb_link'], $ctracker_settings['global_message'], $ctracker_settings['global_message']);
+	$global_message_text = phpbb_stored_text($ctracker_settings['global_message']);
+	$global_message_url = phpbb_profile_http_url($ctracker_settings['global_message']);
+	$global_message_output = ($ctracker_settings['global_message_type'] == 1 || $global_message_url === '')
+		? $global_message_text
+		: sprintf($lang['ctracker_gmb_link'], $global_message_url, $global_message_text);
 	$template->assign_block_vars('ctracker_message', array(
 		'ROW_COLOR' => 'E1FFDF', 'ICON_GLOB' => $images['ctracker_note'],
 		'L_MESSAGE_TEXT' => $global_message_output, 'L_MARK_MESSAGE' => $lang['ctracker_gmb_mark'],